Refactor autoProcessMovie -> movies
