+++ core/__init__.py
@@ -45,7 +45,7 @@ from six.moves import reload_module
 from core import logger, nzbToMediaDB, versionCheck
 from core.auto_process.comics import Comic
 from core.auto_process.autoProcessGames import Game
-from core.auto_process.autoProcessMovie import Movie
+from core.auto_process.movies import Movie
 from core.auto_process.music import Music
 from core.auto_process.tv import TV
 from core.databases import mainDB
