+++ nzbToMedia.py
@@ -631,7 +631,7 @@ import core
 from core import logger, nzbToMediaDB
 from core.auto_process.comics import Comic
 from core.auto_process.autoProcessGames import Game
-from core.auto_process.autoProcessMovie import Movie
+from core.auto_process.movies import Movie
 from core.auto_process.music import Music
 from core.auto_process.tv import TV
 from core.nzbToMediaUserScript import external_script
